requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1
scikit-learn>=1.2.2
shap>=0.41.0
pyyaml>=6.0
lime>=0.2.0
xgboost>=1.7.3  # Optional: for better SHAP support
joblib>=1.1.0
